hygiene(clippy): clear warnings in helpers.rs + scroll_view.rs
- runtime/helpers.rs: remove 6 redundant i32::from() casts (Rect fields
  are already i32 after the u16->i32 Rect refactor in C16)
- widgets/containers/scroll_view.rs: remove needless return statements
  in scroll-action match arms (10 x needless_return); replace .to_vec()
  + .into_iter() with .iter().map() directly (unnecessary_to_owned /
  redundant_iter_cloned)

No behavior change. render.rs and scrollbar.rs excluded per C17 brief
(owned by scrollvis2 workstream).

+++ src/runtime/helpers.rs
@@ -234,10 +234,10 @@ pub fn widget_at_tree_layout(tree: &WidgetTree, x: u16, y: u16) -> Option<NodeId
             render_shift_y -= oy as i32;
         }
         let rect = node.layout_rect;
-        let x0 = i32::from(rect.x0) + render_shift_x;
-        let x1 = i32::from(rect.x1) + render_shift_x;
-        let y0 = i32::from(rect.y0) + render_shift_y;
-        let y1 = i32::from(rect.y1) + render_shift_y;
+        let x0 = rect.x0 + render_shift_x;
+        let x1 = rect.x1 + render_shift_x;
+        let y0 = rect.y0 + render_shift_y;
+        let y1 = rect.y1 + render_shift_y;
         let inside =
             i32::from(x) >= x0 && i32::from(x) < x1 && i32::from(y) >= y0 && i32::from(y) < y1;
         if !inside {
@@ -289,8 +289,8 @@ pub fn tree_content_local_coords(
         render_shift_y -= oy as i32;
     }
 
-    let origin_x = i32::from(rect.x0) + render_shift_x;
-    let origin_y = i32::from(rect.y0) + render_shift_y;
+    let origin_x = rect.x0 + render_shift_x;
+    let origin_y = rect.y0 + render_shift_y;
     let local_x = i32::from(screen_x).saturating_sub(origin_x).max(0) as u16;
     let local_y = i32::from(screen_y).saturating_sub(origin_y).max(0) as u16;
     (local_x, local_y)
